update: set default repo_path to current directory
In `analyze_git_repo_loc.py`, the argparse argument definition for `"repo_path"` has been modified to include a default value of the current directory `"./"`. This allows users to run the script without explicitly specifying the repository path, assuming they are executing the script in the root directory of a Git repository. Additionally, an unused import statement (`from os.path import abspath`) was removed, cleaning up the code dependencies.

This change simplifies the usage for consumers of the script who wish to analyze the repository in which they are currently located without the need to pass the path as an argument.